Add files via upload
The is the first part of a jig to create 10 hard sectored floppies from a 5.25" soft sector floppy.  This is the punch template.  Just align one o the holes with the existing index hole.  Attach the backing place ( separate file/print ).  When locked together, these two pieces will rotate the floppy allowing new holes to be punched through the index hole window.